+++ lib/builtin/theme_picker/theme_picker.dart
@@ -1,2 +1,2 @@
 export 'src/extension.dart';
-export 'src/picker_view.dart';
+export 'src/settings_view.dart';